+++ packages/message-parser/tests/link.test.ts
@@ -264,44 +264,44 @@ test.each([
     ],
   ],
   [
-    '[Github link with hash](https://github.com/QuickSales/Rocket.Chat/pull/26751/files#diff-c87b108ecf1ede549f8ede68eca840fbb330180b927df0b8a0b4df5d06cbd89b)',
+    '[Github link with hash](https://github.com/QuickSales/QuickSales.Vn/pull/26751/files#diff-c87b108ecf1ede549f8ede68eca840fbb330180b927df0b8a0b4df5d06cbd89b)',
     [
       paragraph([
         link(
-          'https://github.com/QuickSales/Rocket.Chat/pull/26751/files#diff-c87b108ecf1ede549f8ede68eca840fbb330180b927df0b8a0b4df5d06cbd89b',
+          'https://github.com/QuickSales/QuickSales.Vn/pull/26751/files#diff-c87b108ecf1ede549f8ede68eca840fbb330180b927df0b8a0b4df5d06cbd89b',
           plain('Github link with hash')
         ),
       ]),
     ],
   ],
   [
-    '[Github link with hash](https://github.com/QuickSales/Rocket.Chat/pull/26751/files#diff)',
+    '[Github link with hash](https://github.com/QuickSales/QuickSales.Vn/pull/26751/files#diff)',
     [
       paragraph([
         link(
-          'https://github.com/QuickSales/Rocket.Chat/pull/26751/files#diff',
+          'https://github.com/QuickSales/QuickSales.Vn/pull/26751/files#diff',
           plain('Github link with hash')
         ),
       ]),
     ],
   ],
   [
-    '[Github link without hash](https://github.com/QuickSales/Rocket.Chat/pull/26751/files)',
+    '[Github link without hash](https://github.com/QuickSales/QuickSales.Vn/pull/26751/files)',
     [
       paragraph([
         link(
-          'https://github.com/QuickSales/Rocket.Chat/pull/26751/files',
+          'https://github.com/QuickSales/QuickSales.Vn/pull/26751/files',
           plain('Github link without hash')
         ),
       ]),
     ],
   ],
   [
-    '[Link with special chars](https://github.com/QuickSales/Rocket.Chat*[/]^_`{}~)',
+    '[Link with special chars](https://github.com/QuickSales/QuickSales.Vn*[/]^_`{}~)',
     [
       paragraph([
         link(
-          'https://github.com/QuickSales/Rocket.Chat*[/]^_`{}~',
+          'https://github.com/QuickSales/QuickSales.Vn*[/]^_`{}~',
           plain('Link with special chars')
         ),
       ]),
@@ -319,10 +319,10 @@ test.each([
     ],
   ],
   [
-    '[Rocket.Chat](https://rocket.chat) Inline Text',
+    '[QuickSales.Vn](https://rocket.chat) Inline Text',
     [
       paragraph([
-        link('https://rocket.chat', plain('Rocket.Chat')),
+        link('https://rocket.chat', plain('QuickSales.Vn')),
         plain(' Inline Text'),
       ]),
     ],
@@ -340,10 +340,10 @@ test.each([
     ],
   ],
   [
-    `[Rocket.Chat](https://rocket.chat)
+    `[QuickSales.Vn](https://rocket.chat)
 Text after in a new line after link`,
     [
-      paragraph([link('https://rocket.chat', plain('Rocket.Chat'))]),
+      paragraph([link('https://rocket.chat', plain('QuickSales.Vn'))]),
       paragraph([plain('Text after in a new line after link')]),
     ],
   ],
@@ -361,11 +361,11 @@ Second line`,
     ],
   ],
   [
-    `[Rocket.Chat](https://rocket.chat)
+    `[QuickSales.Vn](https://rocket.chat)
 
 Text after line break`,
     [
-      paragraph([link('https://rocket.chat', plain('Rocket.Chat'))]),
+      paragraph([link('https://rocket.chat', plain('QuickSales.Vn'))]),
       lineBreak(),
       paragraph([plain('Text after line break')]),
     ],