+++ packages/logo/CHANGELOG.md
@@ -14,4 +14,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-- **logo:** Rocket.Chat logo package ([#535](https://github.com/QuickSales/fuselage/issues/535)) ([94522e6](https://github.com/QuickSales/fuselage/commit/94522e6b74d02f88f56e9ac898ca26b9d1e42fbd))
+- **logo:** QuickSales.Vn logo package ([#535](https://github.com/QuickSales/fuselage/issues/535)) ([94522e6](https://github.com/QuickSales/fuselage/commit/94522e6b74d02f88f56e9ac898ca26b9d1e42fbd))